+++ app/src/main/java/com/example/dentifymobile/navigation/AppContentNavHost.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.example.dentifymobile.dashboard.presentation.navigation.dashboardNavGraph
-import com.example.dentifymobile.patientattention.presentation.navigation.patientAttentionNavGraph
+import com.example.dentifymobile.patientattention.patient.presentation.navigation.patientAttentionNavGraph
 
 @Composable
 fun AppContentNavHost(navController: NavHostController) {
