--- app/src/main/java/com/example/dentifymobile/patientattention/data/model/MedicalHistoryResponse.kt
@@ -1,23 +1,0 @@
-package com.example.dentifymobile.patientattention.data.model
-
-import com.example.dentifymobile.patientattention.domain.model.MedicalHistory
-
-data class MedicalHistoryResponse(
-    val id: Long?,
-    val description: String?,
-    val record: String?,
-    val diagnosis: String?,
-    val medication: String?,
-    val createdAt: String?
-) {
-    fun toDomain(): MedicalHistory {
-        return MedicalHistory(
-            id = id ?: 0L,
-            description = description ?: "",
-            record = record ?: "",
-            diagnosis = diagnosis ?: "",
-            medication = medication ?: "",
-            date = createdAt ?: ""
-        )
-    }
-}
