--- app/src/main/java/com/example/dentifymobile/patientattention/data/remote/dto/MedicalHistoryRequest.kt
@@ -1,8 +1,0 @@
-package com.example.dentifymobile.patientattention.data.remote.dto
-
-data class MedicalHistoryRequest(
-    val description: String,
-    val record: String,
-    val diagnosis: String,
-    val medication: String
-)
