--- app/src/main/java/com/example/dentifymobile/patientattention/data/remote/dto/PatientRequest.kt
@@ -1,10 +1,0 @@
-package com.example.dentifymobile.patientattention.data.remote.dto
-
-data class PatientRequest(
-    val dni: String,
-    val firstName: String,
-    val lastName: String,
-    val email: String,
-    val homeAddress: String,
-    val birthday: String
-)
