--- app/src/main/java/com/example/dentifymobile/patientattention/domain/model/Patient.kt
@@ -1,12 +1,0 @@
-package com.example.dentifymobile.patientattention.domain.model
-
-data class Patient(
-    val id: Long,
-    val dni: String,
-    val firstName: String,
-    val lastName: String,
-    val email: String,
-    val homeAddress: String,
-    val birthday: String,
-    val age: Int
-)
