--- app/src/main/java/com/example/dentifymobile/patientattention/domain/usecases/GetAllMedicalHistoriesUseCase.kt
@@ -1,10 +1,0 @@
-package com.example.dentifymobile.patientattention.domain.usecases
-
-import com.example.dentifymobile.patientattention.domain.model.MedicalHistory
-import com.example.dentifymobile.patientattention.domain.repository.MedicalHistoryRepository
-
-class GetAllMedicalHistoriesUseCase(private val medicalHistoryRepository: MedicalHistoryRepository) {
-    suspend operator fun invoke(id: Long): List<MedicalHistory> {
-        return medicalHistoryRepository.getAllMedicalHistories(id)
-    }
-}
