+++ app/src/main/java/com/example/dentifymobile/patientattention/patient/domain/usecases/GetAllPatientsUseCase.kt
@@ -1,0 +1,15 @@
+package com.example.dentifymobile.patientattention.patient.domain.usecases
+
+
+import com.example.dentifymobile.patientattention.patient.domain.model.Patient
+import com.example.dentifymobile.patientattention.patient.domain.repository.PatientRepository
+
+class GetAllPatientsUseCase(
+    private val repository: PatientRepository
+) {
+
+    suspend operator fun invoke(): List<Patient> {
+        return repository.getAllPatients()
+    }
+
+}
