+++ app/src/main/java/com/example/dentifymobile/patientattention/patient/domain/usecases/UpdatePatientUseCase.kt
@@ -1,0 +1,13 @@
+package com.example.dentifymobile.patientattention.patient.domain.usecases
+
+
+import com.example.dentifymobile.patientattention.patient.domain.model.Patient
+import com.example.dentifymobile.patientattention.patient.domain.repository.PatientRepository
+
+class UpdatePatientUseCase(
+    private val repository: PatientRepository
+) {
+    suspend operator fun invoke(id:Long, patient: Patient) {
+        return repository.updatePatient(id, patient)
+    }
+}
