+++ app/src/main/java/com/example/dentifymobile/patientattention/patient/presentation/viewmodel/PatientsViewModel.kt
@@ -1,0 +1,33 @@
+package com.example.dentifymobile.patientattention.patient.presentation.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+
+import com.example.dentifymobile.patientattention.patient.domain.model.Patient
+import com.example.dentifymobile.patientattention.patient.domain.usecases.DeletePatientUseCase
+import com.example.dentifymobile.patientattention.patient.domain.usecases.GetAllPatientsUseCase
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+
+class PatientsViewModel(
+    private val getAllPatientsUseCase: GetAllPatientsUseCase,
+    private val deletePatientUseCase: DeletePatientUseCase
+): ViewModel() {
+
+    private val _patients = MutableStateFlow<List<Patient>>(emptyList())
+    val patients: StateFlow<List<Patient>> = _patients
+
+    fun getAllPatients() {
+        viewModelScope.launch {
+            _patients.value = getAllPatientsUseCase()
+        }
+    }
+
+    fun deletePatient(id: Long) {
+        viewModelScope.launch {
+            deletePatientUseCase(id)
+            getAllPatients()
+        }
+    }
+}
