--- app/src/main/java/com/example/dentifymobile/patientattention/presentation/viewmodel/MedicalHistoriesViewModel.kt
@@ -1,31 +1,0 @@
-package com.example.dentifymobile.patientattention.presentation.viewmodel
-
-import androidx.compose.runtime.mutableStateOf
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.dentifymobile.patientattention.domain.model.MedicalHistory
-import com.example.dentifymobile.patientattention.domain.usecases.GetAllMedicalHistoriesUseCase
-import com.example.dentifymobile.patientattention.domain.model.Patient
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-
-class MedicalHistoriesViewModel(
-    private val getAllMedicalHistoriesUseCase: GetAllMedicalHistoriesUseCase
-): ViewModel() {
-    private val _medicalHistories = MutableStateFlow<List<MedicalHistory>>(emptyList())
-    val medicalHistories: StateFlow<List<MedicalHistory>> = _medicalHistories
-
-    var selectedPatient = mutableStateOf<Patient?>(null)
-
-    fun getAllMedical(id: Long) {
-        viewModelScope.launch {
-            _medicalHistories.value = getAllMedicalHistoriesUseCase(id)
-        }
-    }
-
-    fun setSelectedPatient(patient: Patient) {
-        selectedPatient.value = patient
-    }
-
-}
